feat: 예매자 type별 필터링 기능 추가
