+++ app/(admin)/admin/layout.tsx
@@ -1,0 +1,9 @@
+'use client';
+import { ReactNode } from 'react';
+import { RecoilRoot } from 'recoil';
+
+const layout = ({ children }: { children: ReactNode }) => {
+  return <RecoilRoot>{children}</RecoilRoot>;
+};
+
+export default layout;
